+++ src/main/java/duke/DukeException.java
@@ -1,4 +1,4 @@
-package main.java;
+package main.java.duke;
 
 public class DukeException extends Exception {
     public DukeException(String message) {
