+++ src/main/java/duke/Parser.java
@@ -1,4 +1,4 @@
-package main.java;
+package main.java.duke;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
