+++ src/main/java/duke/TaskList.java
@@ -1,4 +1,6 @@
-package main.java;
+package main.java.duke;
+
+import main.java.duke.task.*;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
