+++ src/main/java/duke/Ui.java
@@ -1,4 +1,4 @@
-package main.java;
+package main.java.duke;
 
 public class Ui {
     /**
