+++ src/main/java/duke/task/ToDo.java
@@ -1,4 +1,4 @@
-package main.java;
+package main.java.duke.task;
 
 public class ToDo extends Task {
     /**
